Score Exploit Fix
Fixed an issue where users could start any game mode and switch over to
another game mode carrying over their score. This bug has been
squashed. NO CHEATERS!

+++ TapCounter/FastTapperViewController.h
@@ -45,8 +45,8 @@
 @property (strong, nonatomic) IBOutlet UIView *vwSettings;
 
 // Variable Properties
-@property (assign) int clock;
-@property (assign) int gameLength;
+@property (assign) NSUInteger clock;
+@property (assign) NSUInteger gameLength;
 @property (assign) int tapCount;
 @property (assign) float tapsPerSecond;
 @property (assign) BOOL isTimerStarted;
